--- src/App.css
@@ -1,22 +1,0 @@
-@import './index.css';
-
-.container {
-  position: relative;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.button {
-  margin-top: 50px;
-  margin-bottom: 50px;
-  width: 150px;
-  height: 50px;
-  border: none;
-  border-radius: 4px;
-  color: var(--accent-color);
-  background-color: var(--main-color);
-  font-size: 1.8rem;
-}
