+++ src/App.module.css
@@ -1,0 +1,10 @@
+@import './index.css';
+
+.container {
+  position: relative;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
