+++ src/components/Autocomplete/Autocomplete.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { UseFormRegisterReturn } from 'react-hook-form';
+import styles from '../Autocomplete/Autocomplete.module.css';
 
 type Country = {
   name: string;
@@ -12,9 +13,18 @@ type CountriesListProps = {
 
 const DataList: React.FC<CountriesListProps> = ({ list, register }) => {
   return (
-    <div>
-      <label htmlFor="country-choice">Choose a country: </label>
-      <input list="country-list" id="country-choice" {...register} autoComplete="off" />
+    <div className={styles.wrapper}>
+      <label className={styles.label} htmlFor="country-choice">
+        <span className={styles.text}> Choose a country:</span>
+
+        <input
+          list="country-list"
+          id="country-choice"
+          {...register}
+          autoComplete="off"
+          className={styles.input}
+        />
+      </label>
       <datalist id="country-list">
         {list.map((item) => (
           <option value={item.name} key={item.name}></option>
