+++ src/components/Layout.tsx
@@ -1,10 +1,11 @@
 import { Outlet } from 'react-router-dom';
+import styles from '../App.module.css';
 
 const Layout = () => {
   return (
     <>
-      <header>Enter your data, please</header>
-      <main>
+      <header className={styles.header}>Enter your data, please</header>
+      <main className={styles.container}>
         <Outlet />
       </main>
       <footer>Copyright</footer>
