+++ src/pages/UncontrolledForm/UncontrolledForm.module.css
@@ -1,0 +1,63 @@
+.form {
+  width: 400px;
+  padding: 40px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border: 2px solid rgb(30, 4, 30);
+  border-radius: 5px;
+}
+
+.fields {
+}
+
+.wrapper {
+  display: flex;
+  justify-content: space-between;
+}
+
+.gender {
+  text-align: center;
+}
+
+.subtitle {
+  font-size: 1.8rem;
+}
+
+.title {
+  font-size: 2.5rem;
+}
+
+.text {
+  display: block;
+  font-size: 1.5rem;
+  color: black;
+  padding-bottom: 7px;
+}
+
+.image {
+  margin-bottom: 10px;
+}
+.label {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.error {
+  height: 15px;
+  color: brown;
+  font-size: 1rem;
+}
+
+.button {
+  width: 100px;
+  height: 30px;
+  border: none;
+  border-radius: 4px;
+  color: var(--accent-color);
+  background-color: var(--main-color);
+  font-size: 1.6rem;
+}
